+++ clang/tools/libclang/CIndex.cpp
@@ -10181,16 +10181,28 @@ CXString clang_getBinaryOperatorKindSpelling(enum CXBinaryOperatorKind kind) {
 }
 
 enum CXBinaryOperatorKind clang_getCursorBinaryOperatorKind(CXCursor cursor) {
-  if (clang_isExpression(cursor.kind)) {
-    const Expr *expr = getCursorExpr(cursor);
+  if (!clang_isExpression(cursor.kind))
+    return CXBinaryOperator_Invalid;
 
-    if (const auto *op = dyn_cast<BinaryOperator>(expr))
-      return static_cast<CXBinaryOperatorKind>(op->getOpcode() + 1);
+  const Expr *expr = getCursorExpr(cursor);
+  if (!expr)
+    return CXBinaryOperator_Invalid;
 
-    if (const auto *op = dyn_cast<CXXRewrittenBinaryOperator>(expr))
-      return static_cast<CXBinaryOperatorKind>(op->getOpcode() + 1);
+  if (const auto *op = dyn_cast<BinaryOperator>(expr))
+    return static_cast<CXBinaryOperatorKind>(op->getOpcode() + 1);
+
+  if (const auto *op = dyn_cast<CXXRewrittenBinaryOperator>(expr))
+    return static_cast<CXBinaryOperatorKind>(op->getOpcode() + 1);
+
+  if (const auto *OCE = dyn_cast<CXXOperatorCallExpr>(expr)) {
+    const OverloadedOperatorKind Kind = OCE->getOperator();
+    bool isPostfixOp = (OCE->getNumArgs() == 2 &&
+                    (Kind == OO_PlusPlus || Kind == OO_MinusMinus));
+    if (OCE->getNumArgs() == 2 && !isPostfixOp) {
+      auto opcode = BinaryOperator::getOverloadedOpcode(Kind);
+      return static_cast<CXBinaryOperatorKind>(opcode + 1);
+    }
   }
-
   return CXBinaryOperator_Invalid;
 }
 
@@ -10200,11 +10212,27 @@ CXString clang_getUnaryOperatorKindSpelling(enum CXUnaryOperatorKind kind) {
 }
 
 enum CXUnaryOperatorKind clang_getCursorUnaryOperatorKind(CXCursor cursor) {
-  if (clang_isExpression(cursor.kind)) {
-    const Expr *expr = getCursorExpr(cursor);
+  if (!clang_isExpression(cursor.kind)) {
+    return CXUnaryOperator_Invalid;
+  }
 
-    if (const auto *op = dyn_cast<UnaryOperator>(expr))
-      return static_cast<CXUnaryOperatorKind>(op->getOpcode() + 1);
+  const Expr *expr = getCursorExpr(cursor);
+  if (!expr) {
+    return CXUnaryOperator_Invalid;
+  }
+
+  if (const auto *op = dyn_cast<UnaryOperator>(expr)) {
+    return static_cast<CXUnaryOperatorKind>(op->getOpcode() + 1);
+  }
+
+  if (const auto *OCE = dyn_cast<CXXOperatorCallExpr>(expr)) {
+    const OverloadedOperatorKind Kind = OCE->getOperator();
+    bool isPostfixOp = (OCE->getNumArgs() == 2 &&
+                    (Kind == OO_PlusPlus || Kind == OO_MinusMinus));
+    if (OCE->getNumArgs() == 1 || isPostfixOp) {
+      auto opcode = UnaryOperator::getOverloadedOpcode(Kind, isPostfixOp);
+      return static_cast<CXUnaryOperatorKind>(opcode + 1);
+    }
   }
 
   return CXUnaryOperator_Invalid;